Move persistent settings into their own module
The persistent settings were all over the place, so I think it's time to move them all into their own module. A rename to "config" for app persistent settings, as opposed to runtime settings could also be a good idea, but it not part of this commit.

THe InterfaceSettings was also not removed, since it's an odd struct that is tightly integrated with the UI application (for now).
