--- korangar/src/audio/mod.rs
@@ -1,3 +1,0 @@
-mod settings;
-
-pub use self::settings::*;
