+++ korangar/src/graphics/settings.rs
@@ -1,8 +1,9 @@
 use std::fmt::{Display, Formatter};
+#[cfg(feature = "debug")]
+use std::num::NonZeroU32;
 
 #[cfg(feature = "debug")]
-use korangar_debug::logging::{print_debug, Colorize};
-use ron::ser::PrettyConfig;
+use derive_new::new;
 use serde::{Deserialize, Serialize};
 
 use crate::interface::layout::ScreenSize;
@@ -149,67 +150,74 @@ impl Display for ScreenSpaceAntiAliasing {
     }
 }
 
-#[derive(Serialize, Deserialize)]
-pub struct GraphicsSettings {
-    pub vsync: bool,
-    pub limit_framerate: LimitFramerate,
-    pub triple_buffering: bool,
-    pub texture_filtering: TextureSamplerType,
-    pub msaa: Msaa,
-    pub ssaa: Ssaa,
-    pub screen_space_anti_aliasing: ScreenSpaceAntiAliasing,
-    pub shadow_detail: ShadowDetail,
-    pub high_quality_interface: bool,
+#[cfg(feature = "debug")]
+#[derive(Copy, Clone, new)]
+pub struct RenderSettings {
+    #[new(value = "true")]
+    pub show_frames_per_second: bool,
+    #[new(value = "true")]
+    pub frustum_culling: bool,
+    #[new(default)]
+    pub show_bounding_boxes: bool,
+    #[new(value = "true")]
+    pub show_map: bool,
+    #[new(value = "true")]
+    pub show_objects: bool,
+    #[new(value = "true")]
+    pub show_entities: bool,
+    #[new(value = "true")]
+    pub show_water: bool,
+    #[new(value = "true")]
+    pub show_indicators: bool,
+    #[new(value = "true")]
+    pub show_ambient_light: bool,
+    #[new(value = "true")]
+    pub show_directional_light: bool,
+    #[new(value = "true")]
+    pub show_point_lights: bool,
+    #[new(value = "true")]
+    pub show_particle_lights: bool,
+    #[new(default)]
+    pub use_debug_camera: bool,
+    #[new(default)]
+    pub show_wireframe: bool,
+    #[new(default)]
+    pub show_object_markers: bool,
+    #[new(default)]
+    pub show_light_markers: bool,
+    #[new(default)]
+    pub show_sound_markers: bool,
+    #[new(default)]
+    pub show_effect_markers: bool,
+    #[new(default)]
+    pub show_particle_markers: bool,
+    #[new(default)]
+    pub show_entity_markers: bool,
+    #[new(default)]
+    pub show_shadow_markers: bool,
+    #[new(default)]
+    pub show_map_tiles: bool,
+    #[new(default)]
+    pub show_pathing: bool,
+    #[new(default)]
+    pub show_picker_buffer: bool,
+    #[new(default)]
+    pub show_directional_shadow_map: bool,
+    #[new(default)]
+    pub show_point_shadow_map: Option<NonZeroU32>,
+    #[new(default)]
+    pub show_light_culling_count_buffer: bool,
+    #[new(default)]
+    pub show_font_atlas: bool,
 }
 
-impl Default for GraphicsSettings {
-    fn default() -> Self {
-        Self {
-            vsync: true,
-            limit_framerate: LimitFramerate::Unlimited,
-            triple_buffering: true,
-            texture_filtering: TextureSamplerType::Anisotropic(4),
-            msaa: Msaa::X4,
-            ssaa: Ssaa::Off,
-            screen_space_anti_aliasing: ScreenSpaceAntiAliasing::Off,
-            shadow_detail: ShadowDetail::High,
-            high_quality_interface: true,
-        }
-    }
-}
-
-impl GraphicsSettings {
-    const FILE_NAME: &'static str = "client/graphics_settings.ron";
-
-    pub fn new() -> Self {
-        Self::load().unwrap_or_else(|| {
-            #[cfg(feature = "debug")]
-            print_debug!("failed to load graphics settings from {}", Self::FILE_NAME.magenta());
-
-            Default::default()
-        })
-    }
-
-    pub fn load() -> Option<Self> {
-        #[cfg(feature = "debug")]
-        print_debug!("loading graphics settings from {}", Self::FILE_NAME.magenta());
-
-        std::fs::read_to_string(Self::FILE_NAME)
-            .ok()
-            .and_then(|data| ron::from_str(&data).ok())
-    }
-
-    pub fn save(&self) {
-        #[cfg(feature = "debug")]
-        print_debug!("saving graphics settings to {}", Self::FILE_NAME.magenta());
-
-        let data = ron::ser::to_string_pretty(self, PrettyConfig::new()).unwrap();
-        std::fs::write(Self::FILE_NAME, data).expect("unable to write file");
-    }
-}
-
-impl Drop for GraphicsSettings {
-    fn drop(&mut self) {
-        self.save();
+#[cfg(feature = "debug")]
+impl RenderSettings {
+    pub fn show_buffers(&self) -> bool {
+        self.show_directional_shadow_map
+            || self.show_picker_buffer
+            || self.show_point_shadow_map.is_some()
+            || self.show_light_culling_count_buffer
+            || self.show_font_atlas
     }
 }
