+++ korangar/src/settings/mod.rs
@@ -1,0 +1,9 @@
+//! Module that implements all persistent setting files.
+
+mod audio;
+mod graphic;
+mod login;
+
+pub(crate) use audio::*;
+pub(crate) use graphic::*;
+pub(crate) use login::*;
